Move common fixture into conda.testing
